pyproject.toml
[project]
name = "cfa.cloudops"
version = "0.2.9"
description = "Cloud storage, batch, functions, MLOps assistance"
authors = [
    {name = "Ryan Raasch", email = "xng3@cdc.gov"}
]
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "tomli>=2.2.1,<3.0.0",
    "tqdm>=4.67.1,<5.0.0",
    "pandas>=2.3.2",
    "polars>=1.33.1",
    "azure-core>=1.35.1",
    "azure-mgmt-batch>=18.0.0",
    "azure-storage-blob>=12.26.0",
    "azure-identity>=1.25.0",
    "msrest>=0.7.1",
    "msrestazure>=0.6.4.post1",
    "azure-keyvault>=4.2.0",
    "azure-mgmt-compute>=36.0.0",
    "azure-batch>=14.2.0",
    "azure-mgmt-resource>=24.0.0",
    "azure-mgmt-appcontainers>=3.2.0",
    "azure-mgmt-authorization>=4.0.0",
    "python-dotenv>=1.1.1",
    "pathlib>=1.0.0",
    "docker>=7.1.0",
    "pyyaml>=6.0.2",
    "griddler @ git+https://github.com/CDCgov/pygriddler.git",
    "humanize>=4.13.0",
    "toml>=0.10.2",
    "networkx>=3.4.2",
    "anyio>=4.10.0",
    "aiohttp>=3.12.15",
    "openai>=2.11.0"
]

[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["cfa"]

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
hello = "cfa.cloudops.scripts:hello"
create_pool = "cfa.cloudops.scripts:create_pool"
create_job = "cfa.cloudops.scripts:create_job"
create_job_schedule = "cfa.cloudops.scripts:create_job_schedule"
add_task = "cfa.cloudops.scripts:add_task"
create_blob_container = "cfa.cloudops.scripts:create_blob_container"
upload_file = "cfa.cloudops.scripts:upload_file"
upload_folder = "cfa.cloudops.scripts:upload_folder"
monitor_job = "cfa.cloudops.scripts:monitor_job"
check_job_status = "cfa.cloudops.scripts:check_job_status"
delete_job = "cfa.cloudops.scripts:delete_job"
delete_job_schedule = "cfa.cloudops.scripts:delete_job_schedule"
package_and_upload_dockerfile = "cfa.cloudops.scripts:package_and_upload_dockerfile"
upload_docker_image = "cfa.cloudops.scripts:upload_docker_image"
download_file = "cfa.cloudops.scripts:download_file"
download_folder = "cfa.cloudops.scripts:download_folder"
delete_pool = "cfa.cloudops.scripts:delete_pool"
list_blob_files = "cfa.cloudops.scripts:list_blob_files"
delete_blob_file = "cfa.cloudops.scripts:delete_blob_file"
delete_blob_folder = "cfa.cloudops.scripts:delete_blob_folder"
download_job_stats = "cfa.cloudops.scripts:download_job_stats"
download_after_job = "cfa.cloudops.scripts:download_after_job"
add_tasks_from_yaml = "cfa.cloudops.scripts:add_tasks_from_yaml"
generate_sample_env = "cfa.cloudops.scripts:generate_sample_env"
resume_schedule = "cfa.cloudops.scripts:resume_job_schedule"
suspend_schedule = "cfa.cloudops.scripts:suspend_job_schedule"

[tool.pytest.ini_options]
pythonpath = ["."]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.9",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.1.1",
    "coverage-badge>=1.1.2"
]

[tool.poetry.group.mkdocs.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.16"
mkdocstrings = { extras = ["python"], version = "^0.30.0" }
mdx-truly-sane-lists = "^1.3"
